+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added:
 
  + method `get_me` to class `Site`;
+ + method `get_my_inbox` to class `Site`;
  + method `get_my_unread_inbox` to class `Site`;
  + method `post` to class `Site`